Working on the classification; making symbolic links for human curation

+++ cgi/aah-make-review.csh
@@ -20,7 +20,7 @@ if ($?CLOUDANT_URL) then
 else if ($?CN && $?CP) then
     set CU = "$CN":"$CP"@"$CN.cloudant.com"
 else
-    echo `date` "$0 $$ -- no Cloudant URL" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- no Cloudant URL" >>! $TMP/LOG
     goto done
 endif
 
@@ -44,7 +44,7 @@ set OUTPUT = "$TMP/$APP-$API-$QUERY_STRING.$DATE.json"
 set INPROGRESS = ( `echo "$OUTPUT".*` )
 # check OUTPUT in-progress for current interval
 if ($#INPROGRESS) then
-    echo `date` "$0 $$ -- in-progress $DATE" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- in-progress $DATE" >>! $TMP/LOG
     goto done
 else
     onintr done
@@ -56,7 +56,7 @@ if ($DB == "rough-fog" && $?LANIP == 0) then
 else if ($DB == "damp-cloud" && $?LANIP == 0) then
     setenv LANIP "192.168.1.35"
 else
-    echo `date` "$0 $$ -- no LANIP" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- no LANIP" >>! $TMP/LOG
     goto done
 endif
 
@@ -64,18 +64,18 @@ endif
 set old = ( `find "$TMP/" -name "$APP-$API-$QUERY_STRING.*.json" -print | sort -t . -k 2,2 -n -r` )
 if ($#old > 0) then
     set OLD = $old[1]
-    echo `date` "$0 $$ -- found old results ($OLD)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- found old results ($OLD)" >>! $TMP/LOG
     if ($#old > 1) then
-	echo `date` "$0 $$ -- removing old results ($old[2-])" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- removing old results ($old[2-])" >>! $TMP/LOG
 	rm -f $old[2-]
     endif
     # test old
     set class_id = `/usr/local/bin/jq '._id' "$OLD" | sed 's/"//g'`
     if ($class_id != $class) then
-	echo `date` "$0 $$ -- bad results ($OLD)" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- bad results ($OLD)" >>! $TMP/LOG
 	rm -f "$OLD"
     else
-	echo `date` "$0 $$ -- results match ($class_id == $class)" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- results match ($class_id == $class)" >>! $TMP/LOG
     endif
 endif
 
@@ -85,39 +85,39 @@ set prev_seqid = 0
 if ($?OLD == 0) then
     # create temporary file
     set OLD = "$TMP/$APP-$API-$class.$$.json"
-    echo `date` "$0 $$ -- get OLD ($CU/$DB-$API/$class)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- get OLD ($CU/$DB-$API/$class)" >>! $TMP/LOG
     /usr/bin/curl -s -q -o "$OLD" -X GET "$CU/$DB-$API/$class"
     if ($status == 0 && (-s "$OLD")) then
-	echo `date` "$0 $$ -- got OLD ($OLD)" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- got OLD ($OLD)" >>! $TMP/LOG
 	set class_id = `/usr/local/bin/jq '._id' "$OLD" | sed 's/"//g'`
 	if ($class_id != $class) then
-	    echo `date` "$0 $$ -- DB ($class_id) != ID ($class)" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- DB ($class_id) != ID ($class)" >>! $TMP/LOG
 	    rm -f "$OLD"
 	    set prev_seqid = 0
 	else
 	    # get prev_seqid
-	    echo `date` "$0 $$ -- match found ($class_id == $class)" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- match found ($class_id == $class)" >>! $TMP/LOG
 	    # get last sequence for old output
 	    set prev_seqid = `/usr/local/bin/jq '.seqid' "$OLD" | sed 's/"//g'`
 	    if ($status == 0) then
-		echo `date` "$0 $$ -- OLD seqid ($prev_seqid)" >>! $TMP/LOG
+		if ($?DEBUG) echo `date` "$0 $$ -- OLD seqid ($prev_seqid)" >>! $TMP/LOG
 		set date = `/usr/local/bin/jq '.date' "$OLD" | sed 's/"//g'`
 		mv "$OLD" "$TMP/$APP-$API-$QUERY_STRING.$date.json"
 	    else
-		echo `date` "$0 $$ -- invalid OLD ($OLD; $prev_seqid)" >>! $TMP/LOG
+		if ($?DEBUG) echo `date` "$0 $$ -- invalid OLD ($OLD; $prev_seqid)" >>! $TMP/LOG
 	        rm -f "$OLD"
 		set prev_seqid = 0
 	    endif
 	endif
     else
-	echo `date` "$0 $$ ** failure getting $CU/$DB-$API/$class" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ ** failure getting $CU/$DB-$API/$class" >>! $TMP/LOG
 	set prev_seqid = 0
 	rm -f "$OLD"
         goto done
     endif
 endif
 
-echo `date` "$0 $$ -- prev_seqid ($prev_seqid)" >>! $TMP/LOG
+if ($?DEBUG) echo `date` "$0 $$ -- prev_seqid ($prev_seqid)" >>! $TMP/LOG
 
 #
 # get CHANGES records
@@ -125,35 +125,35 @@ echo `date` "$0 $$ -- prev_seqid ($prev_seqid)" >>! $TMP/LOG
 set CHANGES = "$TMP/$APP-$API-$QUERY_STRING-changes.$DATE.json"
 set seqid = $prev_seqid
 if ( ! -e "$CHANGES" ) then
-    echo `date` "$0 $$ -- get changes ($CU/$DB/_changes?descending=true&include_docs=true&since=$prev_seqid" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- get changes ($CU/$DB/_changes?descending=true&include_docs=true&since=$prev_seqid" >>! $TMP/LOG
     /usr/bin/curl -s -q -o "$CHANGES" "$CU/$DB/_changes?descending=true&include_docs=true&since=$prev_seqid" >>&! $TMP/LOG
-    echo `date` "$0 $$ -- got ($CHANGES)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- got ($CHANGES)" >>! $TMP/LOG
     set seqid = ( `/usr/local/bin/jq .last_seq "$CHANGES" | sed 's/"//g'` )
     if ($seqid == "null") then
-         echo `date` "$0 $$ -- failure processing changes" >>! $TMP/LOG
+         if ($?DEBUG) echo `date` "$0 $$ -- failure processing changes" >>! $TMP/LOG
          goto done
     endif
     # remove old changes
     set old = ( `find "$TMP/" -name "$APP-$API-$QUERY_STRING-changes.*.json" -print | sort -t . -k 2,2 -n -r` )
     if ($#old > 1) then
-	echo `date` "$0 $$ -- removing old changes ($old[2-])" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- removing old changes ($old[2-])" >>! $TMP/LOG
 	rm -f $old[2-]
     endif
 else
-    echo `date` "$0 $$ -- changes are current ($TTL) update in " `echo "$SECONDS - $DATE" | bc` >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- changes are current ($TTL) update in " `echo "$SECONDS - $DATE" | bc` >>! $TMP/LOG
 endif
 set seqid = ( `/usr/local/bin/jq .last_seq "$CHANGES" | sed 's/"//g'` )
 if ($#seqid == 0 || $seqid == "null") then
-    echo `date` "$0 $$ -- invalid changes ($CHANGES)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- invalid changes ($CHANGES)" >>! $TMP/LOG
     goto done
 else
-    echo `date` "$0 $$ -- last sequence ($seqid)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- last sequence ($seqid)" >>! $TMP/LOG
 endif
 
 if ($?seqid && $?prev_seqid) then
     if ($#seqid > 0 && $#prev_seqid > 0 && $prev_seqid != 0 && $seqid != 0) then
 	if ("$seqid" == "$prev_seqid") then
-	    echo `date` "$0 $$ !! NO NEW EVENTS ($seqid)" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ !! NO NEW EVENTS ($seqid)" >>! $TMP/LOG
 	    goto done
 	endif
     endif
@@ -163,9 +163,9 @@ set RESULTS = "$TMP/$APP-$API-$QUERY_STRING-results.$DATE.json"
 if (-s "$CHANGES" && (! -s "$RESULTS" || ((-M "$CHANGES") > (-M "$RESULTS")))) then
     # remove old results
     set old = ( `ls -1 "$TMP/$APP-$API-$QUERY_STRING-results".*.json` )
-    echo `date` "$0 $$ -- removing old results ($old)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- removing old results ($old)" >>! $TMP/LOG
     if ($#old > 0) rm -f $old
-    echo `date` "$0 $$ -- creating results from changes" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- creating results from changes" >>! $TMP/LOG
     /usr/local/bin/jq -c '[.results[].doc|{file:.visual.image,tag:.alchemy.text,score:.alchemy.score,year:.year,month:.month,day:.day,hour:.hour,minute:.minute,second:.second}]' "$CHANGES" \
 	| /usr/local/bin/jq -c '{results:.[]|[.file,.tag,.score,.year,.month,.day,.hour,.minute,.second]}' \
 	| sed 's/"//g' \
@@ -177,9 +177,9 @@ if (-s "$CHANGES" && (! -s "$RESULTS" || ((-M "$CHANGES") > (-M "$RESULTS")))) t
 	     printf("{\"file\":\"%s\",\"tag\":\"%s\",\"score\":%f,\"ampm\":\"%s\",\"day\":\"%s\",\"interval\":%d}\n",$1,$2,$3,strftime("%p",t),strftime("%A",t),m); \
 	   }' \
 	| sort -r >! "$RESULTS"
-    echo `date` "$0 $$ -- completed ($RESULTS)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- completed ($RESULTS)" >>! $TMP/LOG
 else
-    echo `date` "$0 $$ -- results are current with changes" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- results are current with changes" >>! $TMP/LOG
 endif
 
 #
@@ -189,12 +189,12 @@ endif
 #
 
 if ($?FTP_GET == 0) then
-    echo `date` "$0 $$ -- getting new images" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- getting new images" >>! $TMP/LOG
     # create temporary output for processing sequentially; ignore "ampm"
     foreach line ( `/usr/local/bin/jq -c '[.file,.tag,.score,.day,.interval]' "$RESULTS" | sed 's/\[//' | sed 's/\]//' | sed 's/ /_/g' | awk -F, '{ printf("%s,%s,%f,%s,%d\n", $1,$2,$3,$4,$5) }'` )
 	set tuple = ( `echo "$line" | sed 's/,/ /g'` )
 	if ($#tuple < 2) then
-	    echo `date` "$0 $$ -- bad tuple ($tuple) ($line)" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- bad tuple ($tuple) ($line)" >>! $TMP/LOG
 	    continue
 	endif
 	set tag = `echo $tuple[2] | sed 's/"//g'`
@@ -210,28 +210,28 @@ if ($?FTP_GET == 0) then
 		set ftp = "ftp://$LANIP/$file" 
 		/usr/bin/curl -s -q "$ftp" -o "$image"
 		if ($status != 0) then
-		    echo `date` "$0 $$ -- fail ($ftp)" >>! $TMP/LOG
+		    if ($?DEBUG) echo `date` "$0 $$ -- fail ($ftp)" >>! $TMP/LOG
 		    break
 		endif
 		if (-s "$image") then
-		    echo `date` "$0 $$ -- success ($image)" >>! $TMP/LOG
+		    if ($?DEBUG) echo `date` "$0 $$ -- success ($image)" >>! $TMP/LOG
 		    # optionally delete the source
 		    if ($?FTP_DELETE) then
-			echo `date` "$0 $$ -- deleting ($file)" >>! $TMP/LOG
+			if ($?DEBUG) echo `date` "$0 $$ -- deleting ($file)" >>! $TMP/LOG
 			/usr/bin/curl -s -q "ftp://$LANIP/" -Q "-DELE $file"
 		    endif
 	        else
-		    echo `date` "$0 $$ -- removing ZERO ($image)" >>! $TMP/LOG
+		    if ($?DEBUG) echo `date` "$0 $$ -- removing ZERO ($image)" >>! $TMP/LOG
 		    rm -f "$image"
 	        endif
 	    else
-		echo `date` "$0 $$ -- done; found existing image ($image)" >>! $TMP/LOG
+		if ($?DEBUG) echo `date` "$0 $$ -- done; found existing image ($image)" >>! $TMP/LOG
 	        break
 	    endif
 	endif
     end
 else
-    echo `date` $0 $$ -- skipping new images" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` $0 $$ -- skipping new images" >>! $TMP/LOG
 endif
 
 #
@@ -240,12 +240,12 @@ endif
 
 # only works for limited # and format of class names (no space)
 set classes = ( `/bin/ls -1 "$TMP/$DB"` )
-echo `date` "$0 $$ -- found $#classes classes" >>! $TMP/LOG
+if ($?DEBUG) echo `date` "$0 $$ -- found $#classes classes" >>! $TMP/LOG
 
 set NEW = "$OUTPUT".$$
 echo -n '{ "seqid":"'$seqid'","date":"'$DATE'","device":"'"$DB"'","count":'$#classes',"classes":[' >! "$NEW"
 
-echo `date` "$0 $$ -- iterating over $classes" >>! $TMP/LOG
+if ($?DEBUG) echo `date` "$0 $$ -- iterating over $classes" >>! $TMP/LOG
 
 @ k = 0
 # this should really be fed by a find(1) command
@@ -253,18 +253,18 @@ foreach i ( $classes )
     if ($k > 0) echo "," >> "$NEW"
     set nfiles = ( `/bin/ls -1 "$TMP/$DB/$i" | wc | awk '{ print $1 }'` )
     echo -n '{"name":"'$i'","count":'$nfiles'}' >> "$NEW"
-    echo `date` "$0 $$ -- $i, $nfiles" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- $i, $nfiles" >>! $TMP/LOG
     @ k++
 end
 echo -n ']}' >> "$NEW"
 
 /usr/local/bin/jq -c '.' "$NEW" >>! $TMP/LOG
 if ($status != 0) then
-    echo `date` "$0 $$ -- malformed JSON: `cat "$NEW"` >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- malformed JSON: `cat "$NEW"` >>! $TMP/LOG
     rm -f "$NEW"
     goto done
 else
-    echo `date` "$0 $$ -- good JSON ($NEW)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- good JSON ($NEW)" >>! $TMP/LOG
 endif
 
 # update statistics
@@ -274,19 +274,19 @@ mv -f "$NEW" "$OUTPUT"
 # update Cloudant
 #
 if ($?CLOUDANT_OFF == 0 && $?CU && $?DB) then
-    echo `date` "$0 $$ -- test if DB exists ($CU/$DB-$API)" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- test if DB exists ($CU/$DB-$API)" >>! $TMP/LOG
     set DEVICE_DB = `/usr/bin/curl -s -q -X GET "$CU/$DB-$API" | /usr/local/bin/jq '.db_name'`
     if ( "$DEVICE_DB" == "null" ) then
-	echo `date` "$0 $$ -- creating DB $CU/$DB-$API" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- creating DB $CU/$DB-$API" >>! $TMP/LOG
         # create DB
         set DEVICE_DB = `/usr/bin/curl -s -q -X PUT "$CU/$DB-$API" | /usr/local/bin/jq '.ok'`
         # test for success
         if ( "$DEVICE_DB" != "true" ) then
             # failure
-	    echo `date` "$0 $$ -- failure creating Cloudant database ($DB-$API)" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- failure creating Cloudant database ($DB-$API)" >>! $TMP/LOG
             setenv CLOUDANT_OFF TRUE
 	else
-	    echo `date` "$0 $$ -- success creating DB $CU/$DB-$API" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- success creating DB $CU/$DB-$API" >>! $TMP/LOG
         endif
     endif
     if ( $?CLOUDANT_OFF == 0 ) then
@@ -295,25 +295,25 @@ if ($?CLOUDANT_OFF == 0 && $?CU && $?DB) then
 	    set doc = ( `cat "$OLD" | /usr/local/bin/jq ._id,._rev | sed 's/"//g'` )
 	    if ($#doc == 2 && $doc[1] == $class && $doc[2] != "") then
 		set rev = $doc[2]
-		echo `date` "$0 $$ -- deleting old output ($rev)" >>! $TMP/LOG
+		if ($?DEBUG) echo `date` "$0 $$ -- deleting old output ($rev)" >>! $TMP/LOG
 		/usr/bin/curl -s -q -X DELETE "$CU/$DB-$API/$class?rev=$rev" >>&! $TMP/LOG
 	    endif
 	else
-            echo `date` "$0 $$ -- no old output to delete" >>! $TMP/LOG
+            if ($?DEBUG) echo `date` "$0 $$ -- no old output to delete" >>! $TMP/LOG
         endif
-        echo `date` "$0 $$ -- storing new output" >>! $TMP/LOG
+        if ($?DEBUG) echo `date` "$0 $$ -- storing new output" >>! $TMP/LOG
         /usr/bin/curl -s -q -H "Content-type: application/json" -X PUT "$CU/$DB-$API/$class" -d "@$OUTPUT" >>&! $TMP/LOG
 	if ($status == 0) then
-	    echo `date` "$0 $$ -- success storing new output" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- success storing new output" >>! $TMP/LOG
 	else
-	    echo `date` "$0 $$ -- failure storing new output" >>! $TMP/LOG
+	    if ($?DEBUG) echo `date` "$0 $$ -- failure storing new output" >>! $TMP/LOG
 	endif
     else
-	echo `date` "$0 $$ -- Cloudant OFF ($DB-$API)" >>! $TMP/LOG
+	if ($?DEBUG) echo `date` "$0 $$ -- Cloudant OFF ($DB-$API)" >>! $TMP/LOG
     endif
 else
-    echo `date` "$0 $$ -- no Cloudant update" >>! $TMP/LOG
+    if ($?DEBUG) echo `date` "$0 $$ -- no Cloudant update" >>! $TMP/LOG
 endif
 
 done:
-echo `date` "$0 $$ -- FINISH"  >>! $TMP/LOG
+echo `date` "$0 $$ -- FINISH ($QUERY_STRING)"  >>! $TMP/LOG
